[EASY-45] refactor: UserService interface 삭제 및 리팩토링
