+++ src/main/java/team/molu/edayserver/service/UserService.java
@@ -1,24 +1,47 @@
 package team.molu.edayserver.service;
 
+import org.springframework.stereotype.Service;
+import reactor.core.publisher.Mono;
 import team.molu.edayserver.domain.Jwt;
 import team.molu.edayserver.domain.Oauth;
 import team.molu.edayserver.domain.User;
+import team.molu.edayserver.exception.UserNotFoundException;
+import team.molu.edayserver.repository.JwtRepository;
+import team.molu.edayserver.repository.OauthRepository;
+import team.molu.edayserver.repository.UserRepository;
 
-public interface UserService {
+@Service
+public class UserService {
+    private final UserRepository userRepository;
+    private final OauthRepository oauthRepository;
+    private final JwtRepository jwtRepository;
+
+    public UserService(UserRepository userRepository, OauthRepository oauthRepository, JwtRepository jwtRepository) {
+        this.userRepository = userRepository;
+        this.oauthRepository = oauthRepository;
+        this.jwtRepository = jwtRepository;
+    }
+
     /**
      * 주어진 email에 해당하는 사용자 정보를 조회합니다.
      *
      * @param email 조회할 사용자 email
      * @return 사용자 객체, 해당 ID의 사용자가 없으면 null 반환
      */
-    User findUserByEmail(String email);
+    public User findUserByEmail(String email) {
+        Mono<User> userMono = userRepository.findUserByEmail(email);
+        return userMono.switchIfEmpty(Mono.error(new UserNotFoundException("User not found with email: " + email)))
+                .block();
+    }
 
     /**
      * 사용자 정보를 저장합니다.
      *
      * @param user 저장할 사용자 객체
      */
-    void createUser(User user);
+    public void createUser(User user) {
+        userRepository.save(user).subscribe();
+    }
 
     /**
      * 사용자 정보를 업데이트합니다.
@@ -26,7 +49,24 @@ public interface UserService {
      * @param user 수정할 사용자 객체
      * @return 정상적으로 수정되었다면 true, 아니라면 false 반환
      */
-    boolean updateUser(User user);
+    public boolean updateUser(User user) {
+        return Boolean.TRUE.equals(userRepository.findUserByEmail(user.getEmail())
+                .switchIfEmpty(Mono.error(new UserNotFoundException("User not found with email: " + user.getEmail())))
+                .flatMap(existingUser -> {
+                    User updatedUser = User.builder()
+                            .id(existingUser.getId())
+                            .email(existingUser.getEmail())
+                            .profileImage(user.getProfileImage() != null ? user.getProfileImage() : existingUser.getProfileImage())
+                            .userRole(existingUser.getUserRole())
+                            .userOauth(existingUser.getUserOauth())
+                            .userJwt(existingUser.getUserJwt())
+                            .build();
+                    return userRepository.save(updatedUser);
+                })
+                .map(savedUser -> true)
+                .onErrorReturn(false)
+                .block());
+    }
 
     /**
      * 사용자를 삭제합니다.
@@ -34,7 +74,14 @@ public interface UserService {
      * @param email 삭제할 사용자 Email
      * @return 정상적으로 삭제되었다면 true, 아니라면 false 반환
      */
-    boolean deleteUserByEmail(String email);
+    public boolean deleteUserByEmail(String email) {
+        return Boolean.TRUE.equals(userRepository.findUserByEmail(email)
+                .switchIfEmpty(Mono.error(new UserNotFoundException("User not found with email: " + email)))
+                .flatMap(existingUser -> userRepository.delete(existingUser)
+                        .then(Mono.just(true)))
+                .onErrorReturn(false)
+                .block());
+    }
 
     /**
      * 사용자 OAuth를 저장합니다.
@@ -42,7 +89,19 @@ public interface UserService {
      * @param oauth 추가할 사용자 관련 oauth
      * @param email 사용자 Email
      */
-    void createUserOauth(Oauth oauth, String email);
+    public void createUserOauth(Oauth oauth, String email) {
+        userRepository.findUserByEmail(email)
+                .switchIfEmpty(Mono.error(new UserNotFoundException("User not found with email: " + email)))
+                .flatMap(user -> {
+                    Oauth newOauth = Oauth.builder()
+                            .oauthId(oauth.getOauthId())
+                            .provider(oauth.getProvider())
+                            .user(user)
+                            .build();
+                    return oauthRepository.save(newOauth);
+                })
+                .block();
+    }
 
     /**
      * 사용자 OAuth를 수정합니다.
@@ -50,7 +109,14 @@ public interface UserService {
      * @param oauth 수정할 사용자 관련 oauth
      * @return 정상적으로 수정되었다면 true, 아니라면 false 반환
      */
-    boolean updateUserOauth(Oauth oauth, String email);
+    public boolean updateUserOauth(Oauth oauth, String email) {
+        return Boolean.TRUE.equals(oauthRepository.findOauthByEmail(email)
+                .switchIfEmpty(Mono.error(new IllegalArgumentException("Oauth not found with email: " + email)))
+                .flatMap(existingOauth -> oauthRepository.save(oauth))
+                .map(savedOauth -> true)
+                .onErrorReturn(false)
+                .block());
+    }
 
     /**
      * 사용자 OAuth를 삭제합니다.
@@ -58,7 +124,14 @@ public interface UserService {
      * @param email 삭제할 사용자 email
      * @return 정상적으로 삭제되었다면 true, 아니라면 false 반환
      */
-    boolean deleteUserOauthByEmail(String email);
+    public boolean deleteUserOauthByEmail(String email) {
+        return Boolean.TRUE.equals(oauthRepository.findOauthByEmail(email)
+                .switchIfEmpty(Mono.error(new IllegalArgumentException("Oauth not found with email: " + email)))
+                .flatMap(existingOauth -> oauthRepository.delete(existingOauth)
+                        .then(Mono.just(true)))
+                .onErrorReturn(false)
+                .block());
+    }
 
     /**
      * 사용자 JWT를 저장합니다.
@@ -66,7 +139,19 @@ public interface UserService {
      * @param jwt 추가할 사용자 관련 jwt
      * @param email 사용자 Email
      */
-    void createUserJwt(Jwt jwt, String email);
+    public void createUserJwt(Jwt jwt, String email) {
+        userRepository.findUserByEmail(email)
+                .switchIfEmpty(Mono.error(new UserNotFoundException("User not found with email: " + email)))
+                .flatMap(user -> {
+                    Jwt newJwt = Jwt.builder()
+                            .refresh(jwt.getRefresh())
+                            .ttl(jwt.getTtl())
+                            .user(user)
+                            .build();
+                    return jwtRepository.save(newJwt);
+                })
+                .block();
+    }
 
     /**
      * 사용자 JWT를 수정합니다.
@@ -74,7 +159,14 @@ public interface UserService {
      * @param jwt 수정할 사용자 관련 jwt
      * @return 정상적으로 수정되었다면 true, 아니라면 false 반환
      */
-    boolean updateUserJwt(Jwt jwt, String email);
+    public boolean updateUserJwt(Jwt jwt, String email) {
+        return Boolean.TRUE.equals(jwtRepository.findJwtByEmail(email)
+                .switchIfEmpty(Mono.error(new IllegalArgumentException("Jwt not found with email: " + email)))
+                .flatMap(existingJwt -> jwtRepository.save(jwt))
+                .map(savedJwt -> true)
+                .onErrorReturn(false)
+                .block());
+    }
 
     /**
      * 사용자 JWT를 삭제합니다.
@@ -82,5 +174,12 @@ public interface UserService {
      * @param email 삭제할 사용자 email
      * @return 정상적으로 삭제되었다면 true, 아니라면 false 반환
      */
-    boolean deleteUserJwtByEmail(String email);
+    public boolean deleteUserJwtByEmail(String email) {
+        return Boolean.TRUE.equals(jwtRepository.findJwtByEmail(email)
+                .switchIfEmpty(Mono.error(new IllegalArgumentException("Jwt not found with email: " + email)))
+                .flatMap(existingJwt -> jwtRepository.delete(existingJwt)
+                        .then(Mono.just(true)))
+                .onErrorReturn(false)
+                .block());
+    }
 }
